Move internal/opts to daemon/internal/opts
