+++ daemon/config/config_linux_test.go
@@ -4,8 +4,8 @@ import (
 	"net/netip"
 	"testing"
 
+	dopts "github.com/docker/docker/daemon/internal/opts"
 	"github.com/docker/docker/daemon/pkg/opts"
-	dopts "github.com/docker/docker/internal/opts"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/moby/moby/api/types/container"
 	"github.com/spf13/pflag"
